feat: split migration page into 2 components for bnUSD and ICX/SODAA
